meeting_planner/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
python-dotenv>=1.0.1

google-api-python-client>=2.130.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0

mcp>=1.1.0

langchain>=0.3.0
langchain-huggingface>=0.1.0
langchain-core>=0.3.0
langgraph>=0.2.0

httpx>=0.27.0